moved ApolloClient to utils
