+++ client/src/App.js
@@ -1,22 +1,10 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import { ApolloProvider } from '@apollo/react-hooks';
-import ApolloClient from 'apollo-boost';
 
+import { ApolloProvider } from '@apollo/react-hooks';
+import client from './utils/Apollo';
 import { Provider } from 'react-redux';
 import store from './utils/State';
-
-const client = new ApolloClient({
-	request: (operation) => {
-		const token = localStorage.getItem('id_token');
-		operation.setContext({
-			headers: {
-				authorization: token ? `Bearer ${token}` : '',
-			},
-		});
-	},
-	uri: '/graphql',
-});
 
 function App() {
 	return (
@@ -24,7 +12,10 @@ function App() {
 			<Provider store={store}>
 				<Router>
 					{/* Header/Nav */}
-					<Switch>{/* <Route exact path={} component={} />*/}</Switch>
+					<Switch>
+						{/* <Route exact path='/' component={Home} /> */}
+						{/* <Route component={NoMatch} /> */}
+					</Switch>
 					{/* Footer */}
 				</Router>
 			</Provider>
