+++ client/src/utils/Apollo.js
@@ -1,0 +1,13 @@
+import ApolloClient from 'apollo-boost';
+
+export default new ApolloClient({
+	request: (operation) => {
+		const token = localStorage.getItem('id_token');
+		operation.setContext({
+			headers: {
+				authorization: token ? `Bearer ${token}` : '',
+			},
+		});
+	},
+	uri: '/graphql',
+});
